LoongArch: Hook up mmwiob() implementation to asm-generic code
LoongArch is a weak ordering architecture, so we once put an completion
memory barrier into arch_spin_unlock(). This is technically correct but
the performance is not so good. We could make use of ARCH_HAS_MMIOWB as
an optimisation to elide the barriers in cases where no I/O writes were
performed inside the critical section.

Thus, we stop overriding arch_spin_unlock(), select ARCH_HAS_MMIOWB and
hook up the asm-generic mmiowb() tracking code for LoongArch, so that an
mmiowb() is automatically issued from spin_unlock() if (and only if) an
I/O write was performed in the critical section.

Signed-off-by: Huacai Chen <chenhuacai@loongson.cn>